Remove event call since events are part of tx struct
